requirements.txt
certifi==2017.11.5
chardet==3.0.4
dj-database-url==0.4.2
Django==1.11.7
django-webpack-loader==0.5.0
djangorestframework==3.7.3
idna==2.6
python-decouple==3.1
pytz==2017.3
requests==2.18.4
urllib3==1.22
whitenoise==3.3.1
waitress
hellowebapp-deploy
psycopg2
